Remove binary files
Git can not diff binary files, so whenever they are changed, git has to
save the entire file. This will slowly increase the size of the
repository. We can not allow this, so we have to make the user compile
their own binaries (or distribute the binaries separately) to play the
game.